+++ Project/src/projekt/fhv/teama/classes/leistungen/IZusatzleistung.java
@@ -8,7 +8,7 @@ public interface IZusatzleistung extends IDao {
 	public void setWarengruppe(IWarengruppe warengruppe);
 	public IWarengruppe getWarengruppe();
 	public void setPreis(float preis);
-	public double getPreis();
+	public float getPreis();
 	public void setExtern(boolean extern);
 	public boolean isExtern();
 }
